Lab4/a2a-agents/requirements.txt
# A2A SDK with HTTP server support (Starlette/ASGI)
a2a-sdk[http-server]>=0.3.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
